extract components to own files
