+++ client/packages/common/src/ui/components/domain/index.ts
@@ -1,2 +1,3 @@
 export * from './Dashboard';
 export * from './StatusCrumbs';
+export * from './SimpleStatistic';
